refactor(macros): rename macros module to macros instead of next
